requirements-web.txt
fastapi
uvicorn[standard]
requests
urllib3
bcrypt
psycopg2-binary
sqlalchemy
python-dotenv
cryptography
pyupbit
python-bithumb
websockets
pydantic
